refactor: test suite cleanup
