+++ package.json
@@ -24,11 +24,11 @@
     "lint:ts:fix": "pnpm run lint:ts --fix",
     "lint:sol": "solhint './packages/contracts/contracts/**/*.sol'",
     "lint:sol:fix": "pnpm run lint:sol --fix --noPrompt",
-    "test:cli": "lerna run test:unit --scope \"maci-cli\"",
     "test:ceremony": "lerna run test:ceremony --scope \"maci-testing\"",
     "test:integration": "lerna run test:integration --scope \"maci-testing\"",
     "test:e2e": "lerna run test:e2e --scope \"maci-testing\"",
-    "test": "lerna run test --ignore maci-cli --ignore maci-testing",
+    "test:unit": "lerna run test:unit --scope \"maci-testing\"",
+    "test": "lerna run test --ignore maci-testing",
     "types": "lerna run types",
     "docs": "lerna run docs",
     "prepare": "is-ci || husky",
