+++ packages/cli/ts/utils/interfaces.ts
@@ -1,5 +1,6 @@
 import type { Signer } from "ethers";
 import type { EContracts } from "maci-sdk";
+
 /**
  * Interface that represents store contract arguments
  */
