+++ packages/testing/.gitignore
@@ -3,4 +3,4 @@ cache
 deploy-config.json
 deployed-contracts.json
 backup
-
+zkeys
